feat: [split-factory-service-named-factory-editable-definition-cluster-002] - [Preserve editable-definition save and validation behavior after the split]
