+++ pkg/service/factory.go
@@ -4,7 +4,6 @@ package service
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
@@ -1588,278 +1587,6 @@ func (fs *FactoryService) SubscribeFactoryEvents(ctx context.Context) (*interfac
 		return nil, fmt.Errorf("subscribe factory events: %w", err)
 	}
 	return stream, nil
-}
-
-// GetEditableFactoryDefinition returns the complete current factory definition
-// with persisted version metadata for graph-editor draft saves.
-func (fs *FactoryService) GetEditableFactoryDefinition(ctx context.Context) (factoryapi.EditableFactoryDefinition, error) {
-	current, err := fs.GetCurrentNamedFactory(ctx)
-	if err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-	version, err := fs.currentFactoryDefinitionVersion(current.Name)
-	if err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-	return factoryapi.EditableFactoryDefinition{
-		FactoryDefinition: current,
-		Version:           version,
-	}, nil
-}
-
-// SaveEditableFactoryDefinition replaces the current named-factory definition
-// with a complete submitted Factory payload and activates the resulting runtime.
-func (fs *FactoryService) SaveEditableFactoryDefinition(ctx context.Context, request factoryapi.SaveEditableFactoryDefinitionRequest) (factoryapi.EditableFactoryDefinition, error) {
-	if fs == nil {
-		return factoryapi.EditableFactoryDefinition{}, fmt.Errorf("factory service is required")
-	}
-	current, payload, err := fs.validateEditableFactorySave(ctx, request)
-	if err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-
-	fs.activationMu.Lock()
-	defer fs.activationMu.Unlock()
-
-	if err := fs.requireIdleRuntime(ctx); err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-	if err := fs.requireFreshEditableFactoryVersion(request.BaseVersion, current.Name); err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-
-	rootDir := fs.factoryRootDir
-	if rootDir == "" && fs.cfg != nil {
-		rootDir = fs.cfg.Dir
-	}
-	factoryDir, err := factoryconfig.ReplaceNamedFactory(rootDir, string(request.FactoryDefinition.Name), payload)
-	if err != nil {
-		switch {
-		case errors.Is(err, factoryconfig.ErrInvalidNamedFactory):
-			return factoryapi.EditableFactoryDefinition{}, fmt.Errorf("%w: %w", ErrInvalidNamedFactory, err)
-		default:
-			return factoryapi.EditableFactoryDefinition{}, err
-		}
-	}
-
-	replacement, err := fs.buildEditableFactoryReplacement(ctx, rootDir, factoryDir)
-	if err != nil {
-		return factoryapi.EditableFactoryDefinition{}, fmt.Errorf("%w: build replacement factory %q: %w", ErrInvalidNamedFactory, request.FactoryDefinition.Name, err)
-	}
-	if err := fs.requireIdleRuntime(ctx); err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-	if err := fs.activateReplacementRuntime(ctx, rootDir, string(request.FactoryDefinition.Name), replacement); err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-
-	saved, err := fs.GetCurrentNamedFactory(ctx)
-	if err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-	version, err := fs.currentFactoryDefinitionVersion(saved.Name)
-	if err != nil {
-		return factoryapi.EditableFactoryDefinition{}, err
-	}
-	return factoryapi.EditableFactoryDefinition{
-		FactoryDefinition: saved,
-		Version:           version,
-	}, nil
-}
-
-func (fs *FactoryService) validateEditableFactorySave(
-	ctx context.Context,
-	request factoryapi.SaveEditableFactoryDefinitionRequest,
-) (factoryapi.Factory, []byte, error) {
-	current, err := fs.GetCurrentNamedFactory(ctx)
-	if err != nil {
-		return factoryapi.Factory{}, nil, err
-	}
-	if current.Name == apisurface.DefaultCurrentFactoryName {
-		return factoryapi.Factory{}, nil, ErrCurrentNamedFactoryNotFound
-	}
-	if request.FactoryDefinition.Name != current.Name {
-		return factoryapi.Factory{}, nil, fmt.Errorf("%w: editable save must preserve current factory name %q", ErrInvalidNamedFactoryName, current.Name)
-	}
-	if err := apisurface.ValidateWritableNamedFactoryName(request.FactoryDefinition.Name); err != nil {
-		return factoryapi.Factory{}, nil, err
-	}
-	if err := validateEditableFactoryTopology(request.FactoryDefinition); err != nil {
-		return factoryapi.Factory{}, nil, err
-	}
-	payload, err := json.Marshal(request.FactoryDefinition)
-	if err != nil {
-		return factoryapi.Factory{}, nil, fmt.Errorf("marshal editable factory payload: %w", err)
-	}
-	return current, payload, nil
-}
-
-func (fs *FactoryService) buildEditableFactoryReplacement(
-	ctx context.Context,
-	rootDir string,
-	factoryDir string,
-) (*replacementFactoryRuntime, error) {
-	sessionID := defaultFactorySessionID
-	if runState := fs.currentRunState(); runState != nil && strings.TrimSpace(runState.sessionID) != "" {
-		sessionID = runState.sessionID
-	}
-	return fs.buildReplacementFactoryRuntime(ctx, rootDir, factoryDir, sessionID)
-}
-
-func (fs *FactoryService) requireFreshEditableFactoryVersion(baseVersion *factoryapi.HybridLogicalTimestamp, name factoryapi.FactoryName) error {
-	rootDir := fs.factoryRootDir
-	if rootDir == "" && fs.cfg != nil {
-		rootDir = fs.cfg.Dir
-	}
-	return fs.requireFreshEditableFactoryVersionAtRoot(baseVersion, rootDir, name)
-}
-
-func (fs *FactoryService) requireFreshEditableFactoryVersionAtRoot(baseVersion *factoryapi.HybridLogicalTimestamp, rootDir string, name factoryapi.FactoryName) error {
-	if baseVersion == nil {
-		return nil
-	}
-	currentVersion, err := fs.currentFactoryDefinitionVersionAtRoot(rootDir, name)
-	if err != nil {
-		return err
-	}
-	if compareEditableFactoryVersions(*baseVersion, currentVersion) < 0 {
-		return fmt.Errorf("%w: base version logical=%d physical=%s current logical=%d physical=%s",
-			apisurface.ErrEditableFactoryVersionStale,
-			baseVersion.Logical,
-			baseVersion.Physical.UTC().Format(time.RFC3339Nano),
-			currentVersion.Logical,
-			currentVersion.Physical.UTC().Format(time.RFC3339Nano),
-		)
-	}
-	return nil
-}
-
-func compareEditableFactoryVersions(left, right factoryapi.HybridLogicalTimestamp) int {
-	if left.Logical < right.Logical {
-		return -1
-	}
-	if left.Logical > right.Logical {
-		return 1
-	}
-	leftPhysical := left.Physical.UTC()
-	rightPhysical := right.Physical.UTC()
-	switch {
-	case leftPhysical.Before(rightPhysical):
-		return -1
-	case leftPhysical.After(rightPhysical):
-		return 1
-	default:
-		return 0
-	}
-}
-
-func validateEditableFactoryTopology(submitted factoryapi.Factory) error {
-	var targets []factoryapi.ErrorTarget
-	targets = append(targets, duplicateNameTargets("workTypes", workTypeNames(submitted.WorkTypes), "node")...)
-	targets = append(targets, duplicateNameTargets("workers", workerNames(submitted.Workers), "node")...)
-	targets = append(targets, duplicateNameTargets("resources", resourceNames(submitted.Resources), "node")...)
-	targets = append(targets, duplicateNameTargets("workstations", workstationNames(submitted.Workstations), "node")...)
-	targets = append(targets, duplicateWorkStateTargets(submitted.WorkTypes)...)
-	targets = append(targets, danglingFactoryReferenceTargets(submitted)...)
-	if len(targets) == 0 {
-		return nil
-	}
-	return apisurface.NewTopologyValidationError("Factory topology contains invalid graph references.", targets)
-}
-
-func duplicateNameTargets(collection string, names []string, kind string) []factoryapi.ErrorTarget {
-	seen := make(map[string]int, len(names))
-	var targets []factoryapi.ErrorTarget
-	for index, name := range names {
-		field := fmt.Sprintf("factoryDefinition.%s[%d].name", collection, index)
-		if strings.TrimSpace(name) == "" {
-			targets = append(targets, editableFactoryErrorTarget("field", "", field))
-			continue
-		}
-		if firstIndex, ok := seen[name]; ok {
-			targets = append(targets,
-				editableFactoryErrorTarget(kind, name, fmt.Sprintf("factoryDefinition.%s[%d].name", collection, firstIndex)),
-				editableFactoryErrorTarget(kind, name, field),
-			)
-			continue
-		}
-		seen[name] = index
-	}
-	return targets
-}
-
-func duplicateWorkStateTargets(workTypes *[]factoryapi.WorkType) []factoryapi.ErrorTarget {
-	if workTypes == nil {
-		return nil
-	}
-	var targets []factoryapi.ErrorTarget
-	for workTypeIndex, workType := range *workTypes {
-		seen := make(map[string]int, len(workType.States))
-		for stateIndex, state := range workType.States {
-			field := fmt.Sprintf("factoryDefinition.workTypes[%d].states[%d].name", workTypeIndex, stateIndex)
-			if strings.TrimSpace(state.Name) == "" {
-				targets = append(targets, editableFactoryErrorTarget("field", workType.Name, field))
-				continue
-			}
-			if firstIndex, ok := seen[state.Name]; ok {
-				id := workType.Name + ":" + state.Name
-				targets = append(targets,
-					editableFactoryErrorTarget("node", id, fmt.Sprintf("factoryDefinition.workTypes[%d].states[%d].name", workTypeIndex, firstIndex)),
-					editableFactoryErrorTarget("node", id, field),
-				)
-				continue
-			}
-			seen[state.Name] = stateIndex
-		}
-	}
-	return targets
-}
-
-func danglingFactoryReferenceTargets(factory factoryapi.Factory) []factoryapi.ErrorTarget {
-	workStates := workStateSet(factory.WorkTypes)
-	workers := stringSet(workerNames(factory.Workers))
-	resources := stringSet(resourceNames(factory.Resources))
-	var targets []factoryapi.ErrorTarget
-	if factory.Workstations == nil {
-		return targets
-	}
-	for workstationIndex, workstation := range *factory.Workstations {
-		if strings.TrimSpace(workstation.Worker) == "" || !workers[workstation.Worker] {
-			targets = append(targets, editableFactoryErrorTarget("field", workstation.Name, fmt.Sprintf("factoryDefinition.workstations[%d].worker", workstationIndex)))
-		}
-		targets = append(targets, danglingIOTargets(workstation.Name, workstation.Inputs, workStates, fmt.Sprintf("factoryDefinition.workstations[%d].inputs", workstationIndex))...)
-		if workstation.Outputs != nil {
-			targets = append(targets, danglingIOTargets(workstation.Name, *workstation.Outputs, workStates, fmt.Sprintf("factoryDefinition.workstations[%d].outputs", workstationIndex))...)
-		}
-		if workstation.OnContinue != nil {
-			targets = append(targets, danglingIOTargets(workstation.Name, *workstation.OnContinue, workStates, fmt.Sprintf("factoryDefinition.workstations[%d].onContinue", workstationIndex))...)
-		}
-		if workstation.OnFailure != nil {
-			targets = append(targets, danglingIOTargets(workstation.Name, *workstation.OnFailure, workStates, fmt.Sprintf("factoryDefinition.workstations[%d].onFailure", workstationIndex))...)
-		}
-		if workstation.OnRejection != nil {
-			targets = append(targets, danglingIOTargets(workstation.Name, *workstation.OnRejection, workStates, fmt.Sprintf("factoryDefinition.workstations[%d].onRejection", workstationIndex))...)
-		}
-		if workstation.Resources != nil {
-			for resourceIndex, resource := range *workstation.Resources {
-				if strings.TrimSpace(resource.Name) == "" || !resources[resource.Name] {
-					targets = append(targets, editableFactoryErrorTarget("edge", workstation.Name+"->"+resource.Name, fmt.Sprintf("factoryDefinition.workstations[%d].resources[%d].name", workstationIndex, resourceIndex)))
-				}
-			}
-		}
-	}
-	return targets
-}
-
-func danglingIOTargets(workstation string, ios []factoryapi.WorkstationIO, workStates map[string]bool, fieldPrefix string) []factoryapi.ErrorTarget {
-	var targets []factoryapi.ErrorTarget
-	for index, io := range ios {
-		id := workstation + "->" + io.WorkType + ":" + io.State
-		if !workStates[io.WorkType+":"+io.State] {
-			targets = append(targets, editableFactoryErrorTarget("edge", id, fmt.Sprintf("%s[%d]", fieldPrefix, index)))
-		}
-	}
-	return targets
 }
 
 func workTypeNames(workTypes *[]factoryapi.WorkType) []string {
